requirements.txt
streamlit>=1.32.0
openai>=1.30.0
pillow>=10.0.0
pdf2image>=1.17.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
weasyprint>=61.0